Fix font-lists being duplicated instead of reusing the same cached storage as intended
This is important in order to keep the same models between Font-List & Specimen.

Note: If the requests are launched at the same time (i.e. on Specimen) the query
is still done twice, this could be prevented by some quirky Promise, but 1/ it wasn't
really the purpose of this fix 2/ this issue would probably not have a major impact
since browser cache could handle this.