Use model-view pattern to keep FontList and Specimen in sync
We're using the font object as a model to be reflected by any views referencing it (i.e. list & specimen).
Instead of triggering the state directly on UI change, we're relaying it to the model which subsequently updates both views state.
This could probably be done better by implementing Flux entirely, however this seems to work fine for now.